Update Device ID to 32 bits in VM's WKPFCOMM, WKPF, and a python dump tool; Debug and correct routing packet handler in master, gateway and device;

+++ src/lib/wkpf/c/arduino/wkpf_config.c
@@ -9,7 +9,7 @@
 
 static char EEMEM eeprom_location[LOCATION_MAX_LENGTH] = ""; // Currently can only handle locations that fit into a single message
 static uint8_t EEMEM eeprom_wkpf_features[WKPF_FEATURE_ARRAY_SIZE];
-static uint8_t EEMEM eeprom_did;
+static uint32_t EEMEM eeprom_did;
 
 #define load_location_length() eeprom_read_byte((uint8_t*)&eeprom_location_length)
 #define save_location_length(x) eeprom_update_byte((uint8_t*)&eeprom_location_length, (uint8_t)x)
